add tools module, implement functions for DropRolls
