+++ src/entities/mod.rs
@@ -21,8 +21,8 @@ pub use self::{
     side_panels::initialize_side_panels,
     spaceship::initialize_spaceship,
     spawn::{
-        spawn_consumable, spawn_drops, spawn_effect, spawn_item, spawn_mob, spawn_random_spawnable,
-        spawn_spawnable,
+        spawn_consumable, spawn_effect, spawn_item, spawn_mob, spawn_spawnable,
+        spawn_spawnable_drop,
     },
     status_bar::initialize_status_bars,
     status_unit::spawn_status_unit,
