+++ src/resources/drops.rs
@@ -1,9 +1,20 @@
-use crate::entities::SpawnableType;
-
+use crate::{
+    entities::{spawn_spawnable_drop, SpawnableType},
+    resources::{
+        ConsumablesResource, EffectsResource, ItemsResource, MobsResource, SpriteSheetsResource,
+    },
+    tools::weighted_rng,
+};
+use amethyst::{
+    core::transform::Transform,
+    ecs::prelude::{Entities, LazyUpdate, ReadExpect},
+};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
-pub type RollProbabilities = Vec<(DropTableType, f32)>;
+pub type DropTablesResource = HashMap<DropTableType, DropProbabilities>;
+pub type DropProbabilities = Vec<(SpawnableType, f32)>;
+pub type RollProbabilities = Vec<(Option<DropTableType>, f32)>;
 
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct DropRolls {
@@ -11,13 +22,50 @@ pub struct DropRolls {
     pub roll_probs: RollProbabilities,
 }
 
+impl DropRolls {
+    fn choose_drop_table(roll_probs: &RollProbabilities) -> &Option<DropTableType> {
+        let probs = roll_probs.iter().map(|roll_prob| roll_prob.1).collect();
+        &roll_probs[weighted_rng(probs)].0
+    }
+
+    fn choose_drop(drop_probs: &DropProbabilities) -> &SpawnableType {
+        let probs = drop_probs.iter().map(|drop_prob| drop_prob.1).collect();
+        &drop_probs[weighted_rng(probs)].0
+    }
+
+    pub fn spawn(
+        &self,
+        spawn_transform: Transform,
+        drop_tables_resource: &ReadExpect<DropTablesResource>,
+        consumables_resource: &ReadExpect<ConsumablesResource>,
+        mobs_resource: &ReadExpect<MobsResource>,
+        items_resource: &ReadExpect<ItemsResource>,
+        effects_resource: &ReadExpect<EffectsResource>,
+        spritesheets_resource: &ReadExpect<SpriteSheetsResource>,
+        entities: &Entities,
+        lazy_update: &ReadExpect<LazyUpdate>,
+    ) {
+        for _ in 0..self.roll_count {
+            // pick a drop table
+            if let Some(drop_table) = Self::choose_drop_table(&self.roll_probs) {
+                spawn_spawnable_drop(
+                    Self::choose_drop(&drop_tables_resource[drop_table]),
+                    spawn_transform.clone(),
+                    consumables_resource,
+                    mobs_resource,
+                    items_resource,
+                    effects_resource,
+                    spritesheets_resource,
+                    entities,
+                    lazy_update,
+                )
+            }
+        }
+    }
+}
+
 #[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
 pub enum DropTableType {
-    NoDrop,
     Standard,
     Boss,
 }
-
-pub type DropProbabilities = Vec<(SpawnableType, f32)>;
-
-pub type DropTablesResource = HashMap<DropTableType, DropProbabilities>;
