+++ src/tools/mod.rs
@@ -1,0 +1,10 @@
+mod math;
+mod random;
+
+pub use self::{
+    math::{
+        distance, dot_product, edge_vector, orthogonal, overlap, poly_to_edges, project, rotate_x,
+        rotate_y, run_sat, sat_is_colliding, signed_modulo, Vector,
+    },
+    random::weighted_rng,
+};
