+++ src/tools/random.rs
@@ -1,0 +1,14 @@
+use rand::{thread_rng, Rng};
+
+pub fn weighted_rng(probs: Vec<f32>) -> usize {
+    let prob_space = probs.iter().fold(0.0, |sum, prob| sum + prob);
+    let pos = thread_rng().gen::<f32>() * prob_space;
+    let mut sum = 0.0;
+    for (idx, prob) in probs.iter().enumerate() {
+        sum += prob;
+        if sum > pos {
+            return idx;
+        }
+    }
+    unreachable!("Error in probabilities.");
+}
